HACK: x86: acpi: Let OS know that console already initialized
Since SPCR 1.04 the default baud rate of 0 means preconfigured UART state.
Because U-Boot does configure console correctly, there is no need
to encode baud rate, parity and amount of stop bits to the OS.

FIXME: There is currently no way to pass the clock frequency or some other
parameters to the OS in case of SPCR and thus it has to be provided by board
code: either through weak method in SPCR generation code, or via a global flag,
or via serial driver... Topic to discuss.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>